+++ src/@types/enum/animal.enum.ts
