+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { TypeormConfig } from './config/typeorm.config';
+import { TypeormConfig } from '@/lib/config/typeorm.config';
 import { MailerModule } from '@nestjs-modules/mailer';
 import { PostsModule } from './domain/posts/posts.module';
 import { CommentsModule } from './domain/comments/comments.module';
@@ -16,13 +16,13 @@ import { PresignUrlsModule } from '@/domain/presign-urls/presign-urls.module';
 import { AuthenticationModule } from './domain/authentication/authentication.module';
 import * as Joi from 'joi';
 import { RedisCacheModule } from './domain/redis-cache/redis-cache.module';
-import { mailerConfigFactory } from './config/mailer.config';
+import { mailerConfigFactory } from '@/lib/config/mailer.config';
 import { RedisModule } from '@nestjs-modules/ioredis';
-import { HealthCheckController } from './health-check/health-check.controller';
+import { HealthCheckController } from '@/domain/health-check/health-check.controller';
 import { ReportsModule } from '@/domain/reports/reports.module';
 import { StickerModule } from '@/domain/sticker/sticker.module';
 import { LikesModule } from '@/domain/likes/likes.module';
-import { SearchModule } from '@/search/search.module';
+import { SearchModule } from '@/domain/search/search.module';
 
 @Module({
   imports: [
