+++ src/domain/items/dto/create-item.dto.ts
@@ -1,5 +1,5 @@
-import { Animal } from '@/domain/types/enum/animal.enum';
-import { ItemType } from '@/domain/types/enum/item-type.enum';
+import { Animal } from '@/@types/enum/animal.enum';
+import { ItemType } from '@/@types/enum/item-type.enum';
 import { ApiProperty } from '@nestjs/swagger';
 import { IsEnum, IsNumber, IsString } from 'class-validator';
