+++ src/lib/exceptions/base.exception.interface.ts
